* tomas/rm-redundant-writes:
  changelog: add #1984
  core/lazy_map+set: avoid calling delete when given key is not present

+++ core/src/ledger/storage_api/collections/lazy_set.rs
@@ -183,7 +183,7 @@ where
         storage.write(&key, ())
     }
 
-    /// Removes a key from the set, returning `true` if the key
+    /// Removes a key from the set if it's present, returning `true` if the key
     /// was in the set.
     pub fn remove<S>(&self, storage: &mut S, key: &K) -> Result<bool>
     where
@@ -191,8 +191,10 @@ where
     {
         let present = self.contains(storage, key)?;
 
-        let key = self.get_key(key);
-        storage.delete(&key)?;
+        if present {
+            let key = self.get_key(key);
+            storage.delete(&key)?;
+        }
 
         Ok(present)
     }
